update dependencies
Added some dependencies. It was still quite chaotic and not consistent. Especially the rgbif dependency: trias referred to the CRAN release, while we need to work with (at least) the github master release. Many package:: removed. Only rgbif:: is left to show easily which functions from rgbif we are using.