Add check limit (100000) if checklist_keys is given
The GBIF API request triggered by name_lookup() cannot return more than 99999 records.  So, an assertion error is added. This is important only when we work with checklist_keys, so that we use name_lookup(). With taxon_keys (via name_usage) there is no problem, because we have to work taxon by taxon.